feat(user): add user with image editing support
update user style

update user with img extension
