+++ src/stores/editor/extensions/User/UserWithImgE.ts
@@ -1,0 +1,8 @@
+import { createUserNodeExtension } from "./createUserNodeExtension.ts";
+
+export function insertUserWithImg() {}
+
+export const UserWithImgExtension = createUserNodeExtension({
+    name: "userWithImg",
+    className: "wj-user-with-img",
+});
